moves over test for schema
